team5/requirements.txt
streamlit
boto3
python-dotenv
mcp
